Package SEO-Analysis for PyPI (installable + CLI, backward compatible)
