+++ src/seo_analysis.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+seo-analysis = seo_analysis.cli:main
