+++ src/seo_analysis.egg-info/requires.txt
@@ -1,0 +1,4 @@
+beautifulsoup4>=4.11
+lxml>=4.9
+openpyxl>=3.0
+requests>=2.28
