+++ src/seo_analysis/__init__.py
@@ -1,0 +1,6 @@
+"""SEO-Analysis: spreadsheet-driven on-page SEO analysis."""
+
+from .analyzer import analyze
+
+__all__ = ["analyze"]
+__version__ = "1.0.0"
